+++ src/jsonSchema.ts
@@ -47,6 +47,7 @@ export interface JSONSchema {
 	const?: any;
 	contains?: JSONSchemaRef;
 	propertyNames?: JSONSchemaRef;
+	examples?: any[];
 
 	// schema draft 07
 	$comment?: string;
@@ -56,7 +57,7 @@ export interface JSONSchema {
 
 	// VSCode extensions
 
-	defaultSnippets?: { label?: string; description?: string; body?: any; bodyText?: string; }[]; // VSCode extension: body: a object that will be converted to a JSON string. bodyText: text with \t and \n
+	defaultSnippets?: { label?: string; description?: string; markdownDescription?: string; body?: any; bodyText?: string; }[]; // VSCode extension: body: a object that will be converted to a JSON string. bodyText: text with \t and \n
 	errorMessage?: string; // VSCode extension
 	patternErrorMessage?: string; // VSCode extension
 	deprecationMessage?: string; // VSCode extension