+++ src/parser/jsonParser.ts
@@ -6,7 +6,7 @@
 
 import * as Json from 'jsonc-parser';
 import { JSONSchema, JSONSchemaRef } from '../jsonSchema';
-import * as objects from '../utils/objects';
+import { isNumber, equals, isBoolean, isString, isDefined } from '../utils/objects';
 import { ASTNode, ObjectASTNode, ArrayASTNode, BooleanASTNode, NumberASTNode, StringASTNode, NullASTNode, PropertyASTNode, JSONPath, ErrorCode } from '../jsonLanguageTypes';
 
 import * as nls from 'vscode-nls';
@@ -144,7 +144,7 @@ export class ObjectASTNodeImpl extends ASTNodeImpl implements ObjectASTNode {
 }
 
 export function asSchema(schema: JSONSchemaRef) {
-	if (typeof schema === 'boolean') {
+	if (isBoolean(schema)) {
 		return schema ? {} : { "not": {} };
 	}
 	return schema;
@@ -168,7 +168,7 @@ export interface ISchemaCollector {
 	schemas: IApplicableSchema[];
 	add(schema: IApplicableSchema): void;
 	merge(other: ISchemaCollector): void;
-	include(node: ASTNode): void;
+	include(node: ASTNode): boolean;
 	newSub(): ISchemaCollector;
 }
 
@@ -224,9 +224,9 @@ export class ValidationResult {
 	}
 
 	public mergeAll(validationResults: ValidationResult[]): void {
-		validationResults.forEach((validationResult) => {
+		for (const validationResult of validationResults) {
 			this.merge(validationResult);
-		});
+		}
 	}
 
 	public merge(validationResult: ValidationResult): void {
@@ -390,9 +390,9 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		}
 		if (Array.isArray(schema.allOf)) {
-			schema.allOf.forEach(subSchemaRef => {
+			for (const subSchemaRef of schema.allOf) {
 				validate(node, asSchema(subSchemaRef), validationResult, matchingSchemas);
-			});
+			}
 		}
 		let notSchema = asSchema(schema.not);
 		if (notSchema) {
@@ -406,10 +406,10 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 					message: localize('notSchemaWarning', "Matches a schema that is not allowed.")
 				});
 			}
-			subMatchingSchemas.schemas.forEach((ms) => {
+			for (const ms of subMatchingSchemas.schemas) {
 				ms.inverted = !ms.inverted;
 				matchingSchemas.add(ms);
-			});
+			}
 		}
 
 		let testAlternatives = (alternatives: JSONSchemaRef[], maxOneMatch: boolean) => {
@@ -417,7 +417,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 
 			// remember the best match that is used for error messages
 			let bestMatch: { schema: JSONSchema; validationResult: ValidationResult; matchingSchemas: ISchemaCollector; } = null;
-			alternatives.forEach(subSchemaRef => {
+			for (const subSchemaRef of alternatives) {
 				let subSchema = asSchema(subSchemaRef);
 				let subValidationResult = new ValidationResult();
 				let subMatchingSchemas = matchingSchemas.newSub();
@@ -445,7 +445,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 						}
 					}
 				}
-			});
+			}
 
 			if (matches.length > 1 && maxOneMatch) {
 				validationResult.problems.push({
@@ -469,12 +469,11 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			testAlternatives(schema.oneOf, true);
 		}
 
-		let testBranch = (schema: JSONSchemaRef) => {
-			let subSchema = asSchema(schema);
+		let testBranch = (schema: JSONSchema) => {
 			let subValidationResult = new ValidationResult();
 			let subMatchingSchemas = matchingSchemas.newSub();
 
-			validate(node, subSchema, subValidationResult, subMatchingSchemas);
+			validate(node, schema, subValidationResult, subMatchingSchemas);
 
 			validationResult.merge(subValidationResult);
 			validationResult.propertiesMatches += subValidationResult.propertiesMatches;
@@ -499,15 +498,16 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		};
 
-		if (schema.if) {
-			testCondition(schema.if, schema.then, schema.else);
+		let ifSchema = asSchema(schema.if);
+		if (ifSchema) {
+			testCondition(ifSchema, asSchema(schema.then), asSchema(schema.else));
 		}
 
 		if (Array.isArray(schema.enum)) {
 			let val = getNodeValue(node);
 			let enumValueMatch = false;
 			for (let e of schema.enum) {
-				if (objects.equals(val, e)) {
+				if (equals(val, e)) {
 					enumValueMatch = true;
 					break;
 				}
@@ -524,9 +524,9 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		}
 
-		if (schema.const) {
+		if (isDefined(schema.const)) {
 			let val = getNodeValue(node);
-			if (!objects.equals(val, schema.const)) {
+			if (!equals(val, schema.const)) {
 				validationResult.problems.push({
 					location: { offset: node.offset, length: node.length },
 					severity: DiagnosticSeverity.Warning,
@@ -554,7 +554,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 	function _validateNumberNode(node: NumberASTNode, schema: JSONSchema, validationResult: ValidationResult, matchingSchemas: ISchemaCollector): void {
 		let val = node.value;
 
-		if (typeof schema.multipleOf === 'number') {
+		if (isNumber(schema.multipleOf)) {
 			if (val % schema.multipleOf !== 0) {
 				validationResult.problems.push({
 					location: { offset: node.offset, length: node.length },
@@ -564,22 +564,22 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		}
 		function getExclusiveLimit(limit: number | undefined, exclusive: boolean | number | undefined): number | undefined {
-			if (typeof exclusive === 'number') {
+			if (isNumber(exclusive)) {
 				return exclusive;
 			}
-			if (typeof exclusive === 'boolean' && exclusive) {
+			if (isBoolean(exclusive) && exclusive) {
 				return limit;
 			}
 			return void 0;
 		}
 		function getLimit(limit: number | undefined, exclusive: boolean | number | undefined): number | undefined {
-			if (typeof exclusive !== 'boolean' || !exclusive) {
+			if (!isBoolean(exclusive) || !exclusive) {
 				return limit;
 			}
 			return void 0;
 		}
 		let exclusiveMinimum = getExclusiveLimit(schema.minimum, schema.exclusiveMinimum);
-		if (typeof exclusiveMinimum === 'number' && val <= exclusiveMinimum) {
+		if (isNumber(exclusiveMinimum) && val <= exclusiveMinimum) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -587,7 +587,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 		let exclusiveMaximum = getExclusiveLimit(schema.maximum, schema.exclusiveMaximum);
-		if (typeof exclusiveMaximum === 'number' && val >= exclusiveMaximum) {
+		if (isNumber(exclusiveMaximum) && val >= exclusiveMaximum) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -595,7 +595,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 		let minimum = getLimit(schema.minimum, schema.exclusiveMinimum);
-		if (typeof minimum === 'number' && val < minimum) {
+		if (isNumber(minimum) && val < minimum) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -603,7 +603,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 		let maximum = getLimit(schema.maximum, schema.exclusiveMaximum);
-		if (typeof maximum === 'number' && val > maximum) {
+		if (isNumber(maximum) && val > maximum) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -613,7 +613,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 	}
 
 	function _validateStringNode(node: StringASTNode, schema: JSONSchema, validationResult: ValidationResult, matchingSchemas: ISchemaCollector): void {
-		if (schema.minLength && node.value.length < schema.minLength) {
+		if (isNumber(schema.minLength) && node.value.length < schema.minLength) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -621,7 +621,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 
-		if (schema.maxLength && node.value.length > schema.maxLength) {
+		if (isNumber(schema.maxLength) && node.value.length > schema.maxLength) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -629,7 +629,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 
-		if (schema.pattern) {
+		if (isString(schema.pattern)) {
 			let regex = new RegExp(schema.pattern);
 			if (!regex.test(node.value)) {
 				validationResult.problems.push({
@@ -694,7 +694,8 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 	function _validateArrayNode(node: ArrayASTNode, schema: JSONSchema, validationResult: ValidationResult, matchingSchemas: ISchemaCollector): void {
 		if (Array.isArray(schema.items)) {
 			let subSchemas = schema.items;
-			subSchemas.forEach((subSchemaRef, index) => {
+			for (let index = 0; index < subSchemas.length; index++) {
+				const subSchemaRef = subSchemas[index];
 				let subSchema = asSchema(subSchemaRef);
 				let itemValidationResult = new ValidationResult();
 				let item = node.items[index];
@@ -704,7 +705,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 				} else if (node.items.length >= subSchemas.length) {
 					validationResult.propertiesValueMatches++;
 				}
-			});
+			}
 			if (node.items.length > subSchemas.length) {
 				if (typeof schema.additionalItems === 'object') {
 					for (let i = subSchemas.length; i < node.items.length; i++) {
@@ -723,11 +724,11 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 		} else {
 			let itemSchema = asSchema(schema.items);
 			if (itemSchema) {
-				node.items.forEach((item) => {
+				for (const item of node.items) {
 					let itemValidationResult = new ValidationResult();
 					validate(item, itemSchema, itemValidationResult, matchingSchemas);
 					validationResult.mergePropertyMatch(itemValidationResult);
-				});
+				}
 			}
 		}
 
@@ -748,7 +749,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		}
 
-		if (schema.minItems && node.items.length < schema.minItems) {
+		if (isNumber(schema.minItems) && node.items.length < schema.minItems) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -756,7 +757,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			});
 		}
 
-		if (schema.maxItems && node.items.length > schema.maxItems) {
+		if (isNumber(schema.maxItems) && node.items.length > schema.maxItems) {
 			validationResult.problems.push({
 				location: { offset: node.offset, length: node.length },
 				severity: DiagnosticSeverity.Warning,
@@ -783,14 +784,14 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 	function _validateObjectNode(node: ObjectASTNode, schema: JSONSchema, validationResult: ValidationResult, matchingSchemas: ISchemaCollector): void {
 		let seenKeys: { [key: string]: ASTNode } = Object.create(null);
 		let unprocessedProperties: string[] = [];
-		node.properties.forEach((node) => {
-			let key = node.keyNode.value;
-			seenKeys[key] = node.valueNode;
+		for (const propertyNode of node.properties) {
+			let key = propertyNode.keyNode.value;
+			seenKeys[key] = propertyNode.valueNode;
 			unprocessedProperties.push(key);
-		});
+		}
 
 		if (Array.isArray(schema.required)) {
-			schema.required.forEach((propertyName: string) => {
+			for (const propertyName of schema.required) {
 				if (!seenKeys[propertyName]) {
 					let keyNode = node.parent && node.parent.type === 'property' && node.parent.keyNode;
 					let location = keyNode ? { offset: keyNode.offset, length: keyNode.length } : { offset: node.offset, length: 1 };
@@ -800,7 +801,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 						message: localize('MissingRequiredPropWarning', 'Missing property "{0}".', propertyName)
 					});
 				}
-			});
+			}
 		}
 
 		let propertyProcessed = (prop: string) => {
@@ -812,12 +813,12 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 		};
 
 		if (schema.properties) {
-			Object.keys(schema.properties).forEach((propertyName: string) => {
+			for (const propertyName of Object.keys(schema.properties)) {
 				propertyProcessed(propertyName);
 				let propertySchema = schema.properties[propertyName];
 				let child = seenKeys[propertyName];
 				if (child) {
-					if (typeof propertySchema === 'boolean') {
+					if (isBoolean(propertySchema)) {
 						if (!propertySchema) {
 							let propertyNode = <PropertyASTNode>child.parent;
 							validationResult.problems.push({
@@ -836,19 +837,19 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 					}
 				}
 
-			});
+			}
 		}
 
 		if (schema.patternProperties) {
-			Object.keys(schema.patternProperties).forEach((propertyPattern: string) => {
+			for (const propertyPattern of Object.keys(schema.patternProperties)) {
 				let regex = new RegExp(propertyPattern);
-				unprocessedProperties.slice(0).forEach((propertyName: string) => {
+				for (const propertyName of unprocessedProperties.slice(0)) {
 					if (regex.test(propertyName)) {
 						propertyProcessed(propertyName);
 						let child = seenKeys[propertyName];
 						if (child) {
 							let propertySchema = schema.patternProperties[propertyPattern];
-							if (typeof propertySchema === 'boolean') {
+							if (isBoolean(propertySchema)) {
 								if (!propertySchema) {
 									let propertyNode = <PropertyASTNode>child.parent;
 									validationResult.problems.push({
@@ -867,22 +868,22 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 							}
 						}
 					}
-				});
-			});
+				}
+			}
 		}
 
 		if (typeof schema.additionalProperties === 'object') {
-			unprocessedProperties.forEach((propertyName: string) => {
+			for (const propertyName of unprocessedProperties) {
 				let child = seenKeys[propertyName];
 				if (child) {
 					let propertyValidationResult = new ValidationResult();
 					validate(child, <any>schema.additionalProperties, propertyValidationResult, matchingSchemas);
 					validationResult.mergePropertyMatch(propertyValidationResult);
 				}
-			});
+			}
 		} else if (schema.additionalProperties === false) {
 			if (unprocessedProperties.length > 0) {
-				unprocessedProperties.forEach((propertyName: string) => {
+				for (const propertyName of unprocessedProperties) {
 					let child = seenKeys[propertyName];
 					if (child) {
 						let propertyNode = <PropertyASTNode>child.parent;
@@ -893,11 +894,11 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 							message: schema.errorMessage || localize('DisallowedExtraPropWarning', 'Property {0} is not allowed.', propertyName)
 						});
 					}
-				});
+				}
 			}
 		}
 
-		if (schema.maxProperties) {
+		if (isNumber(schema.maxProperties)) {
 			if (node.properties.length > schema.maxProperties) {
 				validationResult.problems.push({
 					location: { offset: node.offset, length: node.length },
@@ -907,7 +908,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 			}
 		}
 
-		if (schema.minProperties) {
+		if (isNumber(schema.minProperties)) {
 			if (node.properties.length < schema.minProperties) {
 				validationResult.problems.push({
 					location: { offset: node.offset, length: node.length },
@@ -918,12 +919,12 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 		}
 
 		if (schema.dependencies) {
-			Object.keys(schema.dependencies).forEach((key: string) => {
+			for (const key of Object.keys(schema.dependencies)) {
 				let prop = seenKeys[key];
 				if (prop) {
 					let propertyDep = schema.dependencies[key];
 					if (Array.isArray(propertyDep)) {
-						propertyDep.forEach((requiredProp: string) => {
+						for (const requiredProp of propertyDep) {
 							if (!seenKeys[requiredProp]) {
 								validationResult.problems.push({
 									location: { offset: node.offset, length: node.length },
@@ -933,7 +934,7 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 							} else {
 								validationResult.propertiesValueMatches++;
 							}
-						});
+						}
 					} else {
 						let propertySchema = asSchema(propertyDep);
 						if (propertySchema) {
@@ -943,17 +944,17 @@ function validate(node: ASTNode, schema: JSONSchema, validationResult: Validatio
 						}
 					}
 				}
-			});
+			}
 		}
 
 		let propertyNames = asSchema(schema.propertyNames);
 		if (propertyNames) {
-			node.properties.forEach(f => {
+			for (const f of node.properties) {
 				let key = f.keyNode;
 				if (key) {
 					validate(key, propertyNames, validationResult, NoOpSchemaCollector.instance);
 				}
-			});
+			}
 		}
 	}
 
@@ -1220,7 +1221,7 @@ export function parse(textDocument: TextDocument, config?: JSONDocumentConfig): 
 			let tokenValue = scanner.getTokenValue();
 			try {
 				let numberValue = JSON.parse(tokenValue);
-				if (typeof numberValue !== 'number') {
+				if (!isNumber(numberValue)) {
 					return _error(localize('InvalidNumberFormat', 'Invalid number format.'), ErrorCode.Undefined, node);
 				}
 				node.value = numberValue;