+++ src/services/jsonCompletion.ts
@@ -12,12 +12,13 @@ import { JSONSchema, JSONSchemaRef } from '../jsonSchema';
 import { JSONWorkerContribution, CompletionsCollector } from '../jsonContributions';
 import { stringifyObject } from '../utils/json';
 import { endsWith } from '../utils/strings';
+import { isDefined } from '../utils/objects';
 import {
 	PromiseConstructor, Thenable,
-	ASTNode, ObjectASTNode, ArrayASTNode, PropertyASTNode
+	ASTNode, ObjectASTNode, ArrayASTNode, PropertyASTNode, ClientCapabilities
 } from '../jsonLanguageTypes';
 
-import { CompletionItem, CompletionItemKind, CompletionList, TextDocument, Position, Range, TextEdit, InsertTextFormat } from 'vscode-languageserver-types';
+import { CompletionItem, CompletionItemKind, CompletionList, TextDocument, Position, Range, TextEdit, InsertTextFormat, MarkupContent, MarkupKind } from 'vscode-languageserver-types';
 
 import * as nls from 'vscode-nls';
 const localize = nls.loadMessageBundle();
@@ -25,14 +26,13 @@ const localize = nls.loadMessageBundle();
 
 export class JSONCompletion {
 
-	private schemaService: SchemaService.IJSONSchemaService;
-	private contributions: JSONWorkerContribution[];
-	private promise: PromiseConstructor;
+	private supportsMarkdown: boolean | undefined;
 
-	constructor(schemaService: SchemaService.IJSONSchemaService, contributions: JSONWorkerContribution[] = [], promiseConstructor?: PromiseConstructor) {
-		this.schemaService = schemaService;
-		this.contributions = contributions;
-		this.promise = promiseConstructor || Promise;
+	constructor(
+		private schemaService: SchemaService.IJSONSchemaService,
+		private contributions: JSONWorkerContribution[] = [],
+		private promiseConstructor: PromiseConstructor = Promise,
+		private clientCapabilities: ClientCapabilities = {}) {
 	}
 
 	public doResolve(item: CompletionItem): Thenable<CompletionItem> {
@@ -44,7 +44,7 @@ export class JSONCompletion {
 				}
 			}
 		}
-		return this.promise.resolve(item);
+		return this.promiseConstructor.resolve(item);
 	}
 
 	public doComplete(document: TextDocument, position: Position, doc: Parser.JSONDocument): Thenable<CompletionList> {
@@ -180,7 +180,7 @@ export class JSONCompletion {
 				this.getContributedValueCompletions(doc, node, offset, document, collector, collectionPromises);
 			}
 
-			return this.promise.all(collectionPromises).then(() => {
+			return this.promiseConstructor.all(collectionPromises).then(() => {
 				if (collector.getNumberOfProposals() === 0) {
 					let offsetForSeparator = offset;
 					if (node && (node.type === 'string' || node.type === 'number' || node.type === 'boolean' || node.type === 'null')) {
@@ -209,7 +209,7 @@ export class JSONCompletion {
 								insertText: this.getInsertTextForProperty(key, propertySchema, addValue, separatorAfter),
 								insertTextFormat: InsertTextFormat.Snippet,
 								filterText: this.getFilterTextForValue(key),
-								documentation: propertySchema.description || '',
+								documentation: this.fromMarkup(propertySchema.markdownDescription) || propertySchema.description || '',
 							};
 							if (endsWith(proposal.insertText, `$1${separatorAfter}`)) {
 								proposal.command = {
@@ -469,9 +469,26 @@ export class JSONCompletion {
 				label: this.getLabelForValue(value),
 				insertText: this.getInsertTextForValue(value, separatorAfter),
 				insertTextFormat: InsertTextFormat.Snippet,
-				detail: localize('json.suggest.default', 'Default value'),
+				detail: localize('json.suggest.default', 'Default value')
 			});
 			hasProposals = true;
+		}
+		if (Array.isArray(schema.examples)) {
+			schema.examples.forEach(example => {
+				let type = schema.type;
+				let value = example;
+				for (let i = arrayDepth; i > 0; i--) {
+					value = [value];
+					type = 'array';
+				}
+				collector.add({
+					kind: this.getSuggestionKind(type),
+					label: this.getLabelForValue(value),
+					insertText: this.getInsertTextForValue(value, separatorAfter),
+					insertTextFormat: InsertTextFormat.Snippet
+				});
+				hasProposals = true;
+			});
 		}
 		if (Array.isArray(schema.defaultSnippets)) {
 			schema.defaultSnippets.forEach(s => {
@@ -504,7 +521,7 @@ export class JSONCompletion {
 				collector.add({
 					kind: this.getSuggestionKind(type),
 					label,
-					documentation: s.description,
+					documentation: this.fromMarkup(s.markdownDescription) || s.description,
 					insertText,
 					insertTextFormat: InsertTextFormat.Snippet,
 					filterText
@@ -525,15 +542,17 @@ export class JSONCompletion {
 				label: this.getLabelForValue(schema.const),
 				insertText: this.getInsertTextForValue(schema.const, separatorAfter),
 				insertTextFormat: InsertTextFormat.Snippet,
-				documentation: schema.description
+				documentation: this.fromMarkup(schema.markdownDescription) || schema.description
 			});
 		}
 
 		if (Array.isArray(schema.enum)) {
 			for (let i = 0, length = schema.enum.length; i < length; i++) {
 				let enm = schema.enum[i];
-				let documentation = schema.description;
-				if (schema.enumDescriptions && i < schema.enumDescriptions.length) {
+				let documentation: string | MarkupContent = this.fromMarkup(schema.markdownDescription) || schema.description;
+				if (schema.markdownEnumDescriptions && i < schema.markdownEnumDescriptions.length && this.doesSupportMarkdown()) {
+					documentation = this.fromMarkup(schema.markdownEnumDescriptions[i]);
+				} else if (schema.enumDescriptions && i < schema.enumDescriptions.length) {
 					documentation = schema.enumDescriptions[i];
 				}
 				collector.add({
@@ -645,9 +664,9 @@ export class JSONCompletion {
 	private getInsertTextForValue(value: any, separatorAfter: string): string {
 		var text = JSON.stringify(value, null, '\t');
 		if (text === '{}') {
-			return '{\n\t$1\n}' + separatorAfter;
+			return '{$1}' + separatorAfter;
 		} else if (text === '[]') {
-			return '[\n\t$1\n]' + separatorAfter;
+			return '[$1]' + separatorAfter;
 		}
 		return this.getInsertTextForPlainText(text + separatorAfter);
 	}
@@ -774,10 +793,10 @@ export class JSONCompletion {
 						value = '"$1"';
 						break;
 					case 'object':
-						value = '{\n\t$1\n}';
+						value = '{$1}';
 						break;
 					case 'array':
-						value = '[\n\t$1\n]';
+						value = '[$1]';
 						break;
 					case 'number':
 					case 'integer':
@@ -849,8 +868,23 @@ export class JSONCompletion {
 		}
 		return (token === Json.SyntaxKind.LineCommentTrivia || token === Json.SyntaxKind.BlockCommentTrivia) && scanner.getTokenOffset() <= offset;
 	}
-}
 
-function isDefined(val: any): val is object {
-	return typeof val !== 'undefined';
+	private fromMarkup(markupString: string | undefined): MarkupContent | string | undefined {
+		if (markupString && this.doesSupportMarkdown()) {
+			return {
+				kind: MarkupKind.Markdown,
+				value: markupString
+			};
+		}
+		return undefined;
+	}
+
+	private doesSupportMarkdown() {
+		if (!isDefined(this.supportsMarkdown)) {
+			const completion = this.clientCapabilities.textDocument && this.clientCapabilities.textDocument.completion;
+			this.supportsMarkdown = completion && completion.completionItem && Array.isArray(completion.completionItem.documentationFormat) && completion.completionItem.documentationFormat.indexOf(MarkupKind.Markdown) !== -1;
+		}
+		return this.supportsMarkdown;
+	}
+
 }