+++ src/test/completion.test.ts
@@ -8,14 +8,15 @@ import * as assert from 'assert';
 import * as JsonSchema from '../jsonSchema';
 import * as jsonLanguageService from '../jsonLanguageService';
 
-import { CompletionList, CompletionItemKind, TextDocument, Position } from 'vscode-languageserver-types';
+import { CompletionList, CompletionItemKind, TextDocument, Position, MarkupContent } from 'vscode-languageserver-types';
+import { ClientCapabilities } from '../jsonLanguageTypes';
 
 const applyEdits = TextDocument.applyEdits;
 
 interface ItemDescription {
 	label: string;
 	detail?: string;
-	documentation?: string;
+	documentation?: string | MarkupContent;
 	kind?: CompletionItemKind;
 	resultText?: string;
 	notAvailable?: boolean;
@@ -35,7 +36,7 @@ let assertCompletion = function (completions: CompletionList, expected: ItemDesc
 		assert.equal(match.detail, expected.detail);
 	}
 	if (expected.documentation) {
-		assert.equal(match.documentation, expected.documentation);
+		assert.deepEqual(match.documentation, expected.documentation);
 	}
 	if (expected.kind) {
 		assert.equal(match.kind, expected.kind);
@@ -47,11 +48,11 @@ let assertCompletion = function (completions: CompletionList, expected: ItemDesc
 
 suite('JSON Completion', () => {
 
-	let testCompletionsFor = function (value: string, schema: JsonSchema.JSONSchema, expected: { count?: number, items?: ItemDescription[] }): PromiseLike<void> {
+	let testCompletionsFor = function (value: string, schema: JsonSchema.JSONSchema, expected: { count?: number, items?: ItemDescription[] }, clientCapabilities = ClientCapabilities.LATEST): PromiseLike<void> {
 		let offset = value.indexOf('|');
 		value = value.substr(0, offset) + value.substr(offset + 1);
 
-		let ls = jsonLanguageService.getLanguageService({});
+		let ls = jsonLanguageService.getLanguageService({ clientCapabilities });
 		if (schema) {
 			ls.configure({
 				schemas: [{
@@ -129,7 +130,7 @@ suite('JSON Completion', () => {
 		await testCompletionsFor('[ { "data": { "key": 1, "data": true } }, { "data": |', null, {
 			count: 3,
 			items: [
-				{ label: '{}', resultText: '[ { "data": { "key": 1, "data": true } }, { "data": {\n\t$1\n}' },
+				{ label: '{}', resultText: '[ { "data": { "key": 1, "data": true } }, { "data": {$1}' },
 				{ label: 'true', resultText: '[ { "data": { "key": 1, "data": true } }, { "data": true' },
 				{ label: 'false', resultText: '[ { "data": { "key": 1, "data": true } }, { "data": false' }
 			]
@@ -750,6 +751,29 @@ suite('JSON Completion', () => {
 
 	});
 
+	test('examples', async function () {
+		let schema: JsonSchema.JSONSchema = {
+			type: 'object',
+			properties: {
+				prop: {
+					type: ['string'],
+					examples: ['a', 'b'],
+					default: 'c'
+				}
+			}
+		};
+
+		await testCompletionsFor('{ "prop": | }', schema, {
+			items: [
+				{ label: '"a"', resultText: '{ "prop": "a" }' },
+				{ label: '"b"', resultText: '{ "prop": "b" }' },
+				{ label: '"c"', resultText: '{ "prop": "c" }' }
+			],
+			count: 3
+		});
+
+	});
+
 	test('Const', async function () {
 		let schema: JsonSchema.JSONSchema = {
 			type: 'object',
@@ -840,7 +864,7 @@ suite('JSON Completion', () => {
 		});
 		await testCompletionsFor('|', schema2, {
 			items: [
-				{ label: '{}', resultText: '{\n\t$1\n}' },
+				{ label: '{}', resultText: '{$1}' },
 				{ label: 'def1', documentation: 'def1Desc', resultText: '{\n\t"hello": "${1:world}"\n}' },
 				{ label: '{"hello":["world"]}', resultText: '{\n\t"${1:hello}": [\n\t\t"${2:world}"\n\t]\n}' }
 			]
@@ -899,6 +923,10 @@ suite('JSON Completion', () => {
 					enum: ['e1', 'e2', 'e3'],
 					enumDescriptions: ['E1', 'E2', 'E3'],
 				},
+				'prop2': {
+					description: 'prop2',
+					enum: ['e1', 'e2', 'e3'],
+				},
 			}
 		};
 
@@ -908,6 +936,87 @@ suite('JSON Completion', () => {
 				{ label: '"e2"', documentation: 'E2' }
 			]
 		});
+		await testCompletionsFor('{ "prop2": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: 'prop2' },
+				{ label: '"e2"', documentation: 'prop2' }
+			]
+		});
+	});
+
+	test('Enum markdown description', async function () {
+		let schema: JsonSchema.JSONSchema = {
+			type: 'object',
+			properties: {
+				'prop1': {
+					enum: ['e1', 'e2', 'e3'],
+					markdownEnumDescriptions: ['*E1*', '*E2*', '*E3*'],
+				},
+				'prop2': {
+					enum: ['e1', 'e2', 'e3'],
+					enumDescriptions: ['E1', 'E2', 'E3'],
+					markdownEnumDescriptions: ['*E1*', '*E2*', '*E3*'],
+				},
+				'prop3': {
+					description: 'Hello',
+					markdownDescription: '*Hello*',
+					enum: ['e1', 'e2', 'e3'],
+					markdownEnumDescriptions: ['*E1*', '*E2*', '*E3*'],
+				},
+				'prop4': {
+					markdownDescription: '*prop4*',
+					enum: ['e1', 'e2', 'e3'],
+				},
+				'prop5': {
+					description: 'prop5',
+					markdownDescription: '*prop5*',
+					enum: ['e1', 'e2', 'e3'],
+				},
+			}
+		};
+
+		await testCompletionsFor('{ "prop1": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: { kind: 'markdown', value: '*E1*' } },
+				{ label: '"e2"', documentation: { kind: 'markdown', value: '*E2*' } }
+			]
+		});
+		await testCompletionsFor('{ "prop2": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: { kind: 'markdown', value: '*E1*' } },
+				{ label: '"e2"', documentation: { kind: 'markdown', value: '*E2*' } }
+			]
+		});
+		await testCompletionsFor('{ "prop3": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: { kind: 'markdown', value: '*E1*' } },
+				{ label: '"e2"', documentation: { kind: 'markdown', value: '*E2*' } }
+			]
+		});
+		await testCompletionsFor('{ "prop4": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: { kind: 'markdown', value: '*prop4*' } },
+				{ label: '"e2"', documentation: { kind: 'markdown', value: '*prop4*' } }
+			]
+		});
+		await testCompletionsFor('{ "prop5": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: { kind: 'markdown', value: '*prop5*' } },
+				{ label: '"e2"', documentation: { kind: 'markdown', value: '*prop5*' } }
+			]
+		});
+
+		// without markdown capability
+		await testCompletionsFor('{ "prop1": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: void 0 },
+			]
+		}, {});
+		await testCompletionsFor('{ "prop2": |', schema, {
+			items: [
+				{ label: '"e1"', documentation: 'E1' },
+			]
+		}, {});
 	});
 
 	test('In comment', async function () {
@@ -1034,8 +1143,8 @@ suite('JSON Completion', () => {
 
 		await testCompletionsFor('{ |', schema, {
 			items: [
-				{ label: 'object', resultText: '{ "object": {\n\t$1\n}' },
-				{ label: 'array', resultText: '{ "array": [\n\t$1\n]' },
+				{ label: 'object', resultText: '{ "object": {$1}' },
+				{ label: 'array', resultText: '{ "array": [$1]' },
 				{ label: 'string', resultText: '{ "string": "$1"' },
 				{ label: 'boolean', resultText: '{ "boolean": $1' },
 				{ label: 'oneEnum', resultText: '{ "oneEnum": "${1:foo}"' },